client/src/app/shared/navbar/navbar.component.ts
import { Component } from '@angular/core';
import { NgClass, NgIf } from '@angular/common';
import { RouterLink, RouterLinkActive } from '@angular/router';

@Component({
  selector: 'app-navbar',
  standalone: true,
  imports: [NgClass, NgIf, RouterLink,RouterLinkActive],
  templateUrl: './navbar.component.html',
  styleUrl: './navbar.component.css'
})
export class NavbarComponent {

   isMenuOpen: boolean = false;

   toggleMenu(): void {
     this.isMenuOpen = !this.isMenuOpen;
   }
}
